`helpers` --> `api` to codify backward compatibility and commitments
