+++ tahoe_sites/tests/test_api.py
@@ -5,16 +5,16 @@ import pytest
 from django.conf import settings
 from organizations.models import Organization
 
-from tahoe_sites import helpers
+from tahoe_sites import api
 from tahoe_sites.models import TahoeSiteUUID
 from tahoe_sites.tests.fatories import UserFactory
 from tahoe_sites.tests.test_models import DefaultsForTestsMixin
 from tahoe_sites.tests.utils import create_organization_mapping
 
 
-class TestHelpers(DefaultsForTestsMixin):
+class TestAPIHelpers(DefaultsForTestsMixin):
     """
-    Tests for helpers
+    Tests for API helpers
     """
     def setUp(self):
         super().setUp()
@@ -53,7 +53,7 @@ class TestHelpers(DefaultsForTestsMixin):
         Test get_organization_by_uuid helper when edx-organizations customization is off
         """
         default_site = TahoeSiteUUID.objects.get(organization=self.default_org)
-        assert helpers.get_organization_by_uuid(default_site.site_uuid) == self.default_org
+        assert api.get_organization_by_uuid(default_site.site_uuid) == self.default_org
 
     @pytest.mark.skipif(settings.FEATURES['TAHOE_SITES_USE_ORGS_MODELS'],
                         reason='Runs only when TAHOE_SITES_USE_ORGS_MODELS is off')
@@ -62,7 +62,7 @@ class TestHelpers(DefaultsForTestsMixin):
         Test get_uuid_by_organization helper when edx-organizations customization is off
         """
         default_site = TahoeSiteUUID.objects.get(organization=self.default_org)
-        assert helpers.get_uuid_by_organization(self.default_org) == default_site.site_uuid
+        assert api.get_uuid_by_organization(self.default_org) == default_site.site_uuid
 
     @pytest.mark.skipif(not settings.FEATURES['TAHOE_SITES_USE_ORGS_MODELS'],
                         reason='Runs only when TAHOE_SITES_USE_ORGS_MODELS is on')
@@ -70,7 +70,7 @@ class TestHelpers(DefaultsForTestsMixin):
         """
         Test get_organization_by_uuid helper when edx-organizations customization is on
         """
-        assert helpers.get_organization_by_uuid(self.default_org.edx_uuid) == self.default_org
+        assert api.get_organization_by_uuid(self.default_org.edx_uuid) == self.default_org
 
     @pytest.mark.skipif(not settings.FEATURES['TAHOE_SITES_USE_ORGS_MODELS'],
                         reason='Runs only when TAHOE_SITES_USE_ORGS_MODELS is on')
@@ -78,7 +78,7 @@ class TestHelpers(DefaultsForTestsMixin):
         """
         Test get_uuid_by_organization helper when edx-organizations customization is on
         """
-        assert helpers.get_uuid_by_organization(self.default_org) == self.default_org.edx_uuid
+        assert api.get_uuid_by_organization(self.default_org) == self.default_org.edx_uuid
 
     def test_get_organizations_for_user_default(self):
         """
@@ -87,15 +87,15 @@ class TestHelpers(DefaultsForTestsMixin):
         self._prepare_mapping_data()
 
         # default_user is mapped to 2 of them
-        assert list(helpers.get_organizations_for_user(self.default_user)) == [self.default_org, self.org1]
+        assert list(api.get_organizations_for_user(self.default_user)) == [self.default_org, self.org1]
 
         # user2 is mapped to two of them, one shared with default_user
-        assert list(helpers.get_organizations_for_user(self.user2)) == [self.org1, self.org2]
+        assert list(api.get_organizations_for_user(self.user2)) == [self.org1, self.org2]
 
         # records with inactive user will not be returned
         self.mapping.is_active = False
         self.mapping.save()
-        assert list(helpers.get_organizations_for_user(self.default_user)) == [self.default_org]
+        assert list(api.get_organizations_for_user(self.default_user)) == [self.default_org]
 
     def test_get_organizations_for_user_with_inactive_users(self):
         """
@@ -106,7 +106,7 @@ class TestHelpers(DefaultsForTestsMixin):
 
         self.mapping.is_active = False
         self.mapping.save()
-        assert list(helpers.get_organizations_for_user(self.default_user, with_inactive_users=True)) == [
+        assert list(api.get_organizations_for_user(self.default_user, with_inactive_users=True)) == [
             self.default_org,
             self.org1
         ]
@@ -120,7 +120,7 @@ class TestHelpers(DefaultsForTestsMixin):
 
         self.mapping.is_admin = True
         self.mapping.save()
-        assert list(helpers.get_organizations_for_user(self.default_user, without_admins=True)) == [self.default_org]
+        assert list(api.get_organizations_for_user(self.default_user, without_admins=True)) == [self.default_org]
 
     def test_get_users_of_organization(self):
         """
@@ -129,15 +129,15 @@ class TestHelpers(DefaultsForTestsMixin):
         self._prepare_mapping_data()
 
         # default_org is mapped to default_user
-        assert list(helpers.get_users_of_organization(self.default_org)) == [self.default_user]
+        assert list(api.get_users_of_organization(self.default_org)) == [self.default_user]
 
         # Org1 is mapped to two users
-        assert list(helpers.get_users_of_organization(self.org1)) == [self.default_user, self.user2]
+        assert list(api.get_users_of_organization(self.org1)) == [self.default_user, self.user2]
 
         # inactive users will not be returned
         self.mapping.is_active = False
         self.mapping.save()
-        assert list(helpers.get_users_of_organization(self.org1)) == [self.user2]
+        assert list(api.get_users_of_organization(self.org1)) == [self.user2]
 
     def test_get_users_of_organization_with_inactive_users(self):
         """
@@ -148,7 +148,7 @@ class TestHelpers(DefaultsForTestsMixin):
 
         self.mapping.is_active = False
         self.mapping.save()
-        assert list(helpers.get_users_of_organization(self.org1, with_inactive_users=True)) == [
+        assert list(api.get_users_of_organization(self.org1, with_inactive_users=True)) == [
             self.default_user,
             self.user2
         ]
@@ -162,7 +162,7 @@ class TestHelpers(DefaultsForTestsMixin):
 
         self.mapping.is_admin = True
         self.mapping.save()
-        assert list(helpers.get_users_of_organization(self.org1, without_admins=True)) == [self.user2]
+        assert list(api.get_users_of_organization(self.org1, without_admins=True)) == [self.user2]
 
     def test_is_active_admin_on_organization(self):
         """
@@ -171,8 +171,8 @@ class TestHelpers(DefaultsForTestsMixin):
         """
         self._prepare_mapping_data()
 
-        assert not helpers.is_active_admin_on_organization(user=self.default_user, organization=self.org1)
+        assert not api.is_active_admin_on_organization(user=self.default_user, organization=self.org1)
 
         self.mapping.is_admin = True
         self.mapping.save()
-        assert helpers.is_active_admin_on_organization(user=self.default_user, organization=self.org1)
+        assert api.is_active_admin_on_organization(user=self.default_user, organization=self.org1)
